# spec/openapi.yaml
openapi: 3.1.0
info:
  title: tifl API
  version: 0.0.1
  description: >
    Canonical contract for the tifl API server. The OpenAPI spec is the only
    contract between clients (browser / Tauri / Capacitor) and the server; every
    client is generated or hand-written against it and reaches no server
    internals. Endpoints are added here as their handlers land.
    See context/backend-server.md and context/architecture-overview.md.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /sessions/{id}:
    get:
      summary: Get session detail and persisted generation progress
      description: >
        Returns one tenant-scoped session with the metadata needed to navigate to
        reader/tasks, retry failed generation, or render a progress screen without
        subscribing to SSE. Another user's session returns 404.
      operationId: getSessionDetail
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Hard-delete a session and its owned generated content
      description: >
        Removes the caller's session plus session-owned generated content such as
        generation stages, tasks, phrase sets, story tokens and glossary entries.
        User knowledge and reader exposure history are preserved; if reader
        events reference the generated story, the story row remains as their
        historical anchor.
      operationId: deleteSession
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "204":
          description: Session deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /stories:
    get:
      summary: List the current user's imported (standalone) stories
      description: >
        Newest-first page of the current user's standalone imported stories —
        rows in the stories table with no generation session. Generated
        (session-backed) stories are listed via GET /sessions, not here. The
        title is a clean display label (the "Imported: " topic prefix is
        stripped; untitled imports fall back to the first words of the text).
        By default all languages are included; pass language to scope the page
        to one language.
      operationId: listImportedStories
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: language
          in: query
          required: false
          description: registered language code to scope the page to; omitted lists all languages
          schema:
            type: string
      responses:
        "200":
          description: Imported story page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportedStoryList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /stories/{id}:
    get:
      summary: Load a story for the reader (tokens + knowledge in one call)
      description: >
        Returns the server-tokenized story, authoritative sentence spans, and
        the caller's canonical knowledge map plus per-surface level map for the
        story's language, in a single response. Sentence spans use half-open token
        positions: start_position is included and end_position is one past the
        final token in the sentence. Ordinary reader ratings are keyed to the
        token's exact form_key; explicit lemma/root well-known/ignored actions
        use the canonical knowledge map. Stories are user-scoped; another user's
        story returns 404.
      operationId: getStory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Story tokens + knowledge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryLoad"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a story (imported or generated)
      description: >
        Deletes the owner's story. One endpoint, two meanings: a standalone
        imported story deletes the story row and its tokens/glossary/audio and
        the caller's reader events for it; a generated (session-backed) story
        deletes the whole session via the existing session-delete cascade
        (story, tasks, phrases, generation stages). Ownership is enforced —
        another user's story returns 404.
      operationId: deleteStory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ImportedStory:
      type: object
      description: One row in the imported-story (standalone reader content) list.
      properties:
        story_id: { type: string }
        title:
          type: string
          description: >
            clean display title — the "Imported: " topic prefix is stripped and
            untitled imports fall back to the first words of the text.
        language: { type: string }
        level: { type: string }
        created_at: { type: number, format: double }
      required: [story_id, title, language, level, created_at]
    ImportedStoryList:
      type: object
      properties:
        stories:
          type: array
          items:
            $ref: "#/components/schemas/ImportedStory"
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
      required: [stories, limit, offset, has_more]
...
    SelectedItemCounts:
      type: object
      description: Persisted SelectedItems counts; background items are not currently stored.
      properties:
        targets: { type: integer }
        new: { type: integer }
      required: [targets, new]
    TaskProgress:
      type: object
      properties:
        total: { type: integer }
        completed: { type: integer, description: number of tasks already graded }
        pending: { type: integer }
      required: [total, completed, pending]
    SessionOverview:
      type: object
      description: One home/resume row for the current user.
      properties:
        session_id: { type: string }
        story_id: { type: string, description: "present once story generation has persisted a story" }
        language: { type: string }
        level: { type: string }
        session_type:
          type: string
          enum: [system, topic_guided, expression_guided]
        content_type:
          type: string
          enum: [story, phrase_set]
          description: >
            what the session's content is — a narrative story or a phrase set.
            Derived from the session shape (expression-guided + expression_output
            phrases is a phrase set; everything else is a story). Tells the client
            whether to load /stories/{id} or the phrase set via /sessions/{id}/content.
        topic: { type: string }
        user_expressions:
          type: array
          items: { type: string }
        expression_output:
          type: string
          enum: [phrases, story]
        status:
          type: string
          enum: [pending, generating, ready, reading, complete, failed]
        created_at: { type: number, format: double }
        archived_at: { type: number, format: double }
        reading_started_at: { type: number, format: double }
        completed_at: { type: number, format: double }
        selected_counts:
          $ref: "#/components/schemas/SelectedItemCounts"
        tasks:
          $ref: "#/components/schemas/TaskProgress"
      required:
        - session_id
        - language
        - level
        - session_type
        - content_type
        - status
        - created_at
        - selected_counts
        - tasks
...
    StageSummary:
      type: object
      properties:
        total: { type: integer }
        pending: { type: integer }
        in_progress: { type: integer }
        complete: { type: integer }
        failed: { type: integer }
        active_stage: { type: string }
        failed_stage: { type: string }
      required: [total, pending, in_progress, complete, failed]
    GenerationStageRecord:
      type: object
      properties:
        stage: { type: string }
        status:
          type: string
          enum: [pending, in_progress, complete, failed]
        started_at: { type: number, format: double }
        completed_at: { type: number, format: double }
        error_code: { type: string }
        error_detail: { type: string }
        retry_count: { type: integer }
      required: [stage, status, retry_count]
    SessionDetail:
      allOf:
        - $ref: "#/components/schemas/SessionOverview"
        - type: object
          properties:
            stage_summary:
              $ref: "#/components/schemas/StageSummary"
            stages:
              type: array
              items:
                $ref: "#/components/schemas/GenerationStageRecord"
          required: [stage_summary, stages]
...
    StoryToken:
      type: object
      description: One token of the server-tokenized story; non-word tokens carry no key.
      properties:
        position: { type: integer, description: "0-indexed, stable identifier" }
        surface: { type: string, description: "display form incl. punctuation" }
        key: { type: string, description: "canonical knowledge key; omitted for non-word tokens" }
        surface_key: { type: string, description: "language-owned exact-form rating key; omitted for non-word tokens" }
        form_key: { type: string, description: "opaque key used to look up this form in surface_knowledge" }
        is_word: { type: boolean }
      required: [position, surface, is_word]
    ReaderKnowledge:
      type: object
      description: The reader's canonical per-item view, keyed by lemma/root/stem key.
      properties:
        level:
          type: string
          description: "explicit canonical learner mark; '' (omitted)=none"
          enum: ["", "1", "2", "3", "4", "5", well_known, ignored]
        lookup_count: { type: integer }
      required: [level, lookup_count]
    ReaderSurfaceKnowledge:
      type: object
      description: The reader's per-displayed-form level, keyed by StoryToken.form_key.
      properties:
        level:
          type: string
          description: "learner self-rating for this exact form; ''=unseen"
          enum: ["", "1", "2", "3", "4", "5", well_known, ignored]
      required: [level]
    SentenceSpan:
      type: object
      description: >
        Authoritative sentence boundary over story token positions. Positions
        are half-open: tokens with position >= start_position and < end_position
        are part of the sentence. Text is exactly the sentence text used by the
        sentence-breakdown cache key.
      properties:
        index:
          type: integer
          description: 0-indexed sentence number within the story.
        start_position:
          type: integer
          description: First token position in the sentence, inclusive.
        end_position:
          type: integer
          description: One past the final token position in the sentence.
        text:
          type: string
          description: Readable sentence text reconstructed from story tokens.
      required: [index, start_position, end_position, text]
    StoryLoad:
      type: object
      description: Everything the reader needs at load time, in one response.
      properties:
        story_id: { type: string }
        language: { type: string }
        tokens:
          type: array
          items: { $ref: "#/components/schemas/StoryToken" }
        sentences:
          type: array
          description: Authoritative sentence spans for highlighting and sentence breakdown requests.
          items: { $ref: "#/components/schemas/SentenceSpan" }
        knowledge:
          type: object
          description: "canonical key → canonical knowledge state; unseen items are absent"
          additionalProperties: { $ref: "#/components/schemas/ReaderKnowledge" }
        surface_knowledge:
          type: object
          description: "StoryToken.form_key → exact-form level; unseen forms are absent"
          additionalProperties: { $ref: "#/components/schemas/ReaderSurfaceKnowledge" }
      required: [story_id, language, tokens, sentences, knowledge, surface_knowledge]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"